+++ packages/uikit/src/biz/DateTimePicker/index.tsx
@@ -1,8 +1,8 @@
-import { useMemoizedFn, useUpdateEffect } from 'ahooks'
+import { useMemoizedFn } from 'ahooks'
 import type { Dayjs } from 'dayjs'
-import { useState, useMemo } from 'react'
+import { useMemo, useState } from 'react'
 
-import { useDisclosure } from '../../hooks/index.js'
+import { useDisclosure, useUncontrolled } from '../../hooks/index.js'
 import { IconClock } from '../../icons/index.js'
 import {
   Box,
@@ -22,7 +22,7 @@ import {
 import { dayjs } from '../../utils/dayjs.js'
 import { DEFAULT_TIME_FORMAT } from '../TimeRangePicker/helpers.js'
 
-import { useTimePickerScroll } from './usePickerScroll.js'
+import { TimeScollerPicker, CurrentValueChangedBy } from './TimeScollerPicker.js'
 
 export interface DateTimePickerProps extends Omit<TextInputProps, 'value' | 'onChange' | 'defaultValue'> {
   placeholder?: string
@@ -57,13 +57,33 @@ export const DateTimePicker = ({
   size
 }: DateTimePickerProps) => {
   const [opened, { close, open }] = useDisclosure(false)
-  const [currentValue, setCurrentValue] = useState<Dayjs>(
-    defaultValue ? dayjs(defaultValue) : value ? dayjs(value) : dayjs()
-  )
+  const [currentValue, setCurrentValue] = useUncontrolled({
+    value: value ? dayjs(value) : undefined,
+    defaultValue: defaultValue ? dayjs(defaultValue) : dayjs(),
+    onChange: (v) => {
+      onChange?.(v.toDate())
+    }
+  })
+  const [currentValueChangedBy, setCurrentValueChangedBy] = useState<CurrentValueChangedBy | null>(null)
 
-  const updateCurrentValue = useMemoizedFn((val: Dayjs) => {
-    setCurrentValue(val)
-    onChange?.(val.toDate())
+  const updateCurrentValue = useMemoizedFn((val: Dayjs, from: typeof currentValueChangedBy) => {
+    let next = val
+
+    if (currentValue?.unix() === next.unix()) {
+      return
+    }
+
+    if (startDate && next.valueOf() < startDate.valueOf()) {
+      next = dayjs(startDate)
+    } else if (endDate && next.valueOf() > endDate.valueOf()) {
+      next = dayjs(endDate)
+    }
+
+    setCurrentValue(next)
+    setCurrentValueChangedBy(from)
+    setTimeout(() => {
+      setCurrentValueChangedBy(null)
+    }, 20)
   })
 
   const inputStr = currentValue.format(format)
@@ -74,14 +94,11 @@ export const DateTimePicker = ({
     return `UTC${h >= 0 ? '+' : '-'}${Math.abs(h)}:${m < 10 ? '0' : ''}${m}`
   }, [utcOffset])
 
-  const scroll = useTimePickerScroll(currentValue, opened, startDate, endDate, utcOffset)
-
   const calendarChange = useMemoizedFn((v: Date) => {
     let next = currentValue
     next = next.utcOffset(utcOffset).year(v.getFullYear()).month(v.getMonth()).date(v.getDate())
-    if (currentValue?.unix() !== next.unix()) {
-      updateCurrentValue(next)
-    }
+
+    updateCurrentValue(next, 'calendar')
   })
 
   const timeInputChange = useMemoizedFn((e: React.ChangeEvent<HTMLInputElement>) => {
@@ -91,30 +108,15 @@ export const DateTimePicker = ({
     let next = currentValue
     next = next.utcOffset(utcOffset).hour(v.getHours()).minute(v.getMinutes()).second(v.getSeconds())
 
-    if (startDate && next.valueOf() < startDate.valueOf()) {
-      updateCurrentValue(dayjs(startDate))
-    } else if (endDate && next.valueOf() > endDate.valueOf()) {
-      updateCurrentValue(dayjs(endDate))
-    } else if (currentValue?.unix() !== next.unix()) {
-      updateCurrentValue(next)
-    }
+    updateCurrentValue(next, 'timeInput')
   })
 
-  // user make changes from scroll picker
-  useUpdateEffect(() => {
+  const timeScrollPickerChange = useMemoizedFn((v: [number, number, number]) => {
+    const [h, m, s] = v
     let next = currentValue
-    const [h, m, s] = scroll.value
     next = next.utcOffset(utcOffset).hour(h).minute(m).second(s)
-    if (currentValue?.unix() !== next.unix()) {
-      updateCurrentValue(next)
-    }
-  }, [scroll.value])
-
-  useUpdateEffect(() => {
-    if (value && dayjs(value).unix() !== currentValue.unix()) {
-      updateCurrentValue(dayjs(value))
-    }
-  }, [value])
+    updateCurrentValue(next, 'timeScroller')
+  })
 
   return (
     <Popover position="bottom-end" opened={opened} withinPortal={withinPortal} shadow="md" onClose={close}>
@@ -220,9 +222,14 @@ export const DateTimePicker = ({
                     sx={(theme) => ({ zIndex: -1, borderRadius: theme.defaultRadius, pointerEvents: 'none' })}
                   />
                 </Flex>
-                <Flex mah="100%" gap={8}>
-                  {scroll.content}
-                </Flex>
+                <TimeScollerPicker
+                  currentValue={currentValue}
+                  currentValueChangedBy={currentValueChangedBy}
+                  onChange={timeScrollPickerChange}
+                  start={startDate}
+                  end={endDate}
+                  utcOffset={utcOffset}
+                />
               </Box>
             </Stack>
           </Group>