+++ packages/documentation/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @tidbcloud/uikit-documentation
 
+## 0.1.30-beta.83
+
+### Patch Changes
+
+- Updated dependencies
+  - @tidbcloud/uikit@2.0.0-beta.83
+
 ## 0.1.30-beta.82
 
 ### Patch Changes